Fix AppVeyor build
Using MSYS2 as MinGW's gcc is too old and doesn't know the
`-fno-sized-deallocation` flag yet.

Public domain declaration for my (Boris-Chengbiao Zhou) contributions:

I dedicate any and all copyright interest in my contributions to this
project to the public domain. I make this dedication for the benefit of
the public at large and to the detriment of my heirs and successors. I
intend this dedication to be an overt act of relinquishment in
perpetuity of all present and future rights to this software under
copyright law.